+++ version_test.go
@@ -10,7 +10,7 @@ import (
 const (
 	Expected_VersionMajor uint16 = 0
 	Expected_VersionMinor uint16 = 1
-	Expected_VersionPatch uint16 = 0
+	Expected_VersionPatch uint16 = 1
 	Expected_VersionAB    uint16 = 0xFFFF
 )
 
@@ -22,9 +22,9 @@ func Test_Version_Elements(t *testing.T) {
 }
 
 func Test_Version(t *testing.T) {
-	require.Equal(t, uint64(0x0000_0001_0000_FFFF), to_be.Version)
+	require.Equal(t, uint64(0x0000_0001_0001_FFFF), to_be.Version)
 }
 
 func Test_Version_String(t *testing.T) {
-	require.Equal(t, "0.1.0", to_be.VersionString())
+	require.Equal(t, "0.1.1", to_be.VersionString())
 }