+++ CHANGES.md
@@ -1,5 +1,11 @@
 # **to-be.Go** Changes
 
+## 0.1.1 - 12th August 2025
+
+* added **examples/libver**;
+* fleshed out **examples/to0**;
+
+
 ## 0.1.0 - 12th August 2025
 
 * added `StringIsFalsey()`, `StringIsTruey()`, `StringIsTruthy()`;